pyproject.toml
[project]
name = "llm-evaluations"
version = "0.1.0"
description = "LLM evaluations on various financial research tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.58.2",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
